requirements.txt
psycopg2-binary
sqlalchemy
geopandas
pymongo==4.10.1
protobuf==4.25.0
intake
jinja2
requests
python-dotenv
intake-sqlite
pyodbc
fastapi
uvicorn
grpcio
grpcio-tools
faker
plotly
apyori
ZODB
matplotlib